+++ @navikt/core/css/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/ds-css
 
+## 1.5.2
+
+### Patch Changes
+
+- :memo: Rename FilterChips to ToggleChips
+
 ## 1.5.1
 
 ### Patch Changes