+++ @navikt/internal/css/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @navikt/ds-css-internal
 
+## 1.5.2
+
 ## 1.5.1
 
 ## 1.5.0